requirements.txt
Django==4.2.20
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2 